+++ R/scripts.R
@@ -1,35 +1,49 @@
 #' Convert SummarizedExperiment or Dataframe to Matrix
 #'
-#' This internal function converts SummarizedExperiment objects and dataframes
-#' (both S3 and S4) to matrices of expression values. Used within oncomix
+#' This function converts SummarizedExperiment objects and dataframes
+#' (both S3 and S4) to matrices of expression values. Used within receptLoss
 #' functions to convert all matrix-like objects to the matrix class.
 #'
 #' @param m Can be a matrix, a data.frame, a DataFrame, or
 #' SummarizedExperiment object.
+#' @param rwnms the rownames of the object. If NA (the default),
+#' assumes that the matrix-like object already has rownames, which
+#' in this case do not need to be supplied separately.
 #' @keywords internal
 #' @importFrom SummarizedExperiment assay
 #' @return A matrix of expression values
 #' @keywords internal
 #' @export
 #' @examples
-#' m <- as.data.frame(matrix(data=rgamma(n=150, shape=2, rate=2),
-#' nrow=10, ncol=15))
+#' m <- as.data.frame(matrix(data=rgamma(n=100, shape=3, rate=2),
+#' nrow=10, ncol=10))
 #' m <- toMatrix(m)
 
-toMatrix <- function(m){
-    if(is.matrix(m)){
-        return(m)
+toMatrix <- function(m, rwnms=NA){
+    ## if an object of class matrix is provided to 'toMatrix',
+    ## the object will be returned without modification
+    ## unless rwnms is specified
+
+    ## check to make sure length of rwnms equals
+    ## number of rows in object (if rwnms is not NA)
+    if(!is.na(rwnms)){
+        if(length(rwnms) != nrow(m)){
+        stop('the length of rwnms should be the same as the # of rows in m')
+        }
     }
+
     if(class(m
-            ) == "SummarizedExperiment"){
+            ) %in% c("SummarizedExperiment", "RangedSummarizedExperiment")){
         m <- assay(m)
-        return(m)
     }
     if(is.data.frame(m) | class(m
             ) == "DataFrame"){
         m <- as.matrix(m)
-        return(m)
     }
+    if(!is.na(rwnms)){
+        rownames(m) <- rwnms
+    }
+    return(m)
 }